Crate rome_service

Re-exports

pub use crate::configuration::create_config;

pub use crate::configuration::load_config;

pub use crate::configuration::Configuration;

pub use crate::configuration::ConfigurationError;

pub use crate::configuration::RuleConfiguration;

pub use crate::matcher::MatchOptions;

pub use crate::matcher::Matcher;

pub use crate::matcher::Pattern;

pub use crate::workspace::Workspace;

Modules

configuration

This module contains the configuration of rome.json

matcher

settings

workspace

The Workspace is the main entry point for high-level clients (the Rome CLI and Language Server) into the various language-specific services of the Rome toolchain (parser, formatter, analyzer …)

Structs

App

This is the main entrypoint of the application.

JsFormatSettings

Rules

Enums

DynRef

Clone of std::borrow::Cow specialized for storing a trait object and holding a mutable reference in the Borrowed variant instead of requiring the inner type to implement std::borrow::ToOwned

RomeError

Generic errors thrown during rome operations

TransportError

Error emitted by the underlying transport layer for a remote Workspace

WorkspaceRef

Constants

VERSION