Function rome_service::configuration::to_analyzer_configuration

pub fn to_analyzer_configuration<ToGlobals>(
    linter_settings: &LinterSettings,
    language_settings: &LanguagesSettings,
    to_globals: ToGlobals
) -> AnalyzerConfiguration where
    ToGlobals: FnOnce(&LanguagesSettings) -> Vec<String>, 

Converts a [WorkspaceSettings] into a suited configuration for the analyzer.

The function needs access to a filter, in order to have an easy access to the metadata of the rules.

The third argument is a closure that accepts a reference to linter_settings.

The closure is responsible to map the globals from the correct location of the settings.

Examples

use rome_service::configuration::to_analyzer_configuration;
use rome_service::settings::{LanguagesSettings, WorkspaceSettings};
let mut settings = WorkspaceSettings::default();
settings.languages.javascript.globals = Some(["jQuery".to_string(), "React".to_string()].into());
// map globals from JS language
let analyzer_configuration =
    to_analyzer_configuration(&settings.linter, &settings.languages, |settings| {
        if let Some(globals) = settings.javascript.globals.as_ref() {
            globals
                .iter()
                .map(|global| global.to_string())
                .collect::<Vec<_>>()
        } else {
            vec![]
        }
    });

 assert_eq!(
    analyzer_configuration.globals,
    vec!["jQuery".to_string(), "React".to_string()]
 )