Trait rome_service::workspace::Workspace

pub trait Workspace: Send + Sync + RefUnwindSafe {
    fn supports_feature(&self, params: SupportsFeatureParams) -> bool;
    fn update_settings(
        &self, 
        params: UpdateSettingsParams
    ) -> Result<(), RomeError>;
    fn open_file(&self, params: OpenFileParams) -> Result<(), RomeError>;
    fn get_syntax_tree(
        &self, 
        params: GetSyntaxTreeParams
    ) -> Result<String, RomeError>;
    fn change_file(&self, params: ChangeFileParams) -> Result<(), RomeError>;
    fn close_file(&self, params: CloseFileParams) -> Result<(), RomeError>;
    fn pull_diagnostics(
        &self, 
        params: PullDiagnosticsParams
    ) -> Result<Vec<Diagnostic>, RomeError>;
    fn pull_actions(
        &self, 
        params: PullActionsParams
    ) -> Result<Vec<AnalyzerAction<JsLanguage>>, RomeError>;
    fn format_file(
        &self, 
        params: FormatFileParams
    ) -> Result<Printed, RomeError>;
    fn format_range(
        &self, 
        params: FormatRangeParams
    ) -> Result<Printed, RomeError>;
    fn format_on_type(
        &self, 
        params: FormatOnTypeParams
    ) -> Result<Printed, RomeError>;
    fn fix_file(
        &self, 
        params: FixFileParams
    ) -> Result<FixFileResult, RomeError>;
}

Required Methods

fn supports_feature(&self, params: SupportsFeatureParams) -> bool

Checks whether a certain feature is supported. There are different conditions:

Rome doesn’t recognize a file, so it can provide the feature;

the feature is disabled inside the configuration;

fn update_settings(&self, params: UpdateSettingsParams) -> Result<(), RomeError>

Update the global settings for this workspace

fn open_file(&self, params: OpenFileParams) -> Result<(), RomeError>

Add a new file to the workspace

fn get_syntax_tree(
&self,
params: GetSyntaxTreeParams
) -> Result<String, RomeError>

fn change_file(&self, params: ChangeFileParams) -> Result<(), RomeError>

Change the content of an open file

fn close_file(&self, params: CloseFileParams) -> Result<(), RomeError>

Remove a file from the workspace

fn pull_diagnostics(
&self,
params: PullDiagnosticsParams
) -> Result<Vec<Diagnostic>, RomeError>

Retrieves the list of diagnostics associated to a file

fn pull_actions(
&self,
params: PullActionsParams
) -> Result<Vec<AnalyzerAction<JsLanguage>>, RomeError>

Retrieves the list of code actions available for a given cursor position within a file

fn format_file(&self, params: FormatFileParams) -> Result<Printed, RomeError>

Runs the given file through the formatter using the provided options and returns the resulting source code

fn format_range(&self, params: FormatRangeParams) -> Result<Printed, RomeError>

Runs a range of an open document through the formatter

fn format_on_type(
&self,
params: FormatOnTypeParams
) -> Result<Printed, RomeError>

Runs a “block” ending at the specified character of an open document through the formatter

fn fix_file(&self, params: FixFileParams) -> Result<FixFileResult, RomeError>

Return the content of the file with all safe code actions applied